pyproject.toml
[tool.poetry]
name = "Cordevall-Python"
version = "0.1.0"
description = "A python Discord Bot"
authors = []

[tool.poetry.dependencies]
python = ">=3.7"
discord = "*"
pyfiglet = "*"
termcolor = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"